+++ kermit/k95/kui/ikui.h
@@ -52,8 +52,8 @@ struct _kui_init {
     int  noclose;
 };
 
-void KuiSetProperty( int propid, long param1, long param2 );
-void KuiGetProperty( int propid, long param1, long param2 );
+void KuiSetProperty( int propid, intptr_t param1, intptr_t param2 );
+void KuiGetProperty( int propid, intptr_t param1, intptr_t param2 );
 
 int KuiInit( void* hInstance, struct _kui_init * );
 HANDLE KuiThreadInit( void* hInstance );
